DTS: Description: 统一db 表字段变更
Team:
Feature or Bugfix:

Match-id-3ae26b55c21116597548f60a997e16d736dcddcc

+++ test/msprof_cpp/analysis_ut/viewer/test/hbm_processor_utest.cpp
@@ -127,7 +127,7 @@ TEST_F(HBMProcessorUTest, TestRunShouldReturnFalseWhenInsertDataFailed)
 
 TEST_F(HBMProcessorUTest, TestRunShouldReturnFalseWhenReserveFailedThenDataIsEmpty)
 {
-    using TempT = std::tuple<uint16_t, uint64_t, uint64_t, uint8_t, uint16_t>;
+    using TempT = std::tuple<uint16_t, uint64_t, uint64_t, uint8_t, uint64_t>;
     MOCKER_CPP(&std::vector<TempT>::reserve).stubs().will(throws(std::bad_alloc()));
     auto processor = HBMProcessor(DB_PATH, PROF_PATHS);
     EXPECT_FALSE(processor.Run());
